Revert "sci-geosciences/osm-gps-map: add ebuild for version 1.1.0"
Unfortunately in spite of the upstream statement that 0.7.3 and 1.0.0+
can be installed in parallel there are several file collisions between
the two versions, some of which (notably introspection support) are not
trivial to fix on our end. Filing an upstream bug report, hopefully the
next release will not require jumping through so many hoops.

This reverts commit f84d5f38689d291273eab2a3d735f07711eecde9.