dev-perl/Cache: Normalise variable order
This ebuild had a very unusual variable order where KEYWORDS
was buried at the bottom.

This makes any tooling that relies on determining keyword
and dependency matches in a single pass more complicated and harder
to be efficient, because it means you have to read the whole
file in every case just in case keywords are declared after dependencies.

Tooling that makes this mistake is fixable, but a more consistent
standard of ebuild also helps a lot.

Package-Manager: portage-2.3.0