net-irc/unrealircd: Precreate tune/logfile for bug #560790.
If you run a program as root and it creates files those will be
owned by root instead of the intended user. Since unrealircd was
not intended to be run directly as root but it is easy to do so
by accident, this change precreates some files so that it is easier
to switch back to running unrealircd as an unprivileged user.

Package-Manager: portage-2.2.28