media-sound/rplay: new revision with EAPI=6.
The original purpose of this new revision was to fix the use of
einstall reported in bug 521564. However, einstall proved necessary,
so along with the update to EAPI=6 -- where einstall is banned -- I
inlined its behavior.

The multilib and eutils eclasses, now unused, were dropped. The latest
debian patchset was also included.

Package-Manager: portage-2.2.28