net-dialup/capifwd: new revision sans eutils.eclass.
Now that the latest revision uses EAPI=6, the eutils eclass (no longer
needed) has been dropped. The autotools eclass looked like it could go
too, but the eautoreconf is actually present to fix a QA warning (not
because we patch e.g. configure.ac) so I added a comment to that
effect.

To make PATCHES work, I revisioned the two patches to be -p1
compatible. The src_prepare and src_install phases now take advantage
of "default" and are a tiny bit simpler.

Package-Manager: portage-2.2.28