games-board/stockfish: new version 7 that respects $CC; remove old ones.
There is a new version of this that compiles fine with the existing
ebuilds. But the build system for this package has a whitelist of
compilers that makes it hard to get it to respect $CC. Fortunately,
there's a hack in the makefile to allow the package to build with
Travis CI, and we can hijack it to force our compiler of choice to be
used.

The version 6 ebuilds have been removed, since they were never
stabilized.

Gentoo-Bug: 542932

Package-Manager: portage-2.2.28