www-client/seamonkey-2.42.3.0_p0: build seamonkey-2.42 out of thunderbird-45 distfiles
Seamonkey upstream is having large difficulties in rolling releases, so
its time to take matters into our own hands.  Packages suffixed with a _p0 version
are packages that will be built out of thunderbird's distfiles, with the added
extensions that are not normally included in the thunderbird codebase.

Package-Manager: portage-2.2.28