dev-libs/beignet: new package.
An OpenCL implementation for Intel GPUs.

Multilib-aware, tested with on amd64 with ABI_X86="32 64". Builds with both
stable versions of sys-devel/llvm currently in Portage (3.5.0 and 3.7.1-r3).

Tests provided by upstream are not run because they require permissions
to access the GPU which the Portage user does not have.

Gentoo-Bug: 545924
Package-Manager: portage-2.3.0