www-servers/lighttpd: Add live ebuild
The situation with lighttpd is rough; upstream has suggested we wait
for 1.4.42 and stabilize it instead. With 1.4.40 and 1.4.41 having
known bugs and the extant buggy stable versions, I'm adding a live
ebuild to give users a practical solution to their bugs while we wait
for a good version to settle on.

lighttpd-1.4.42 is due in October.

Package-Manager: portage-2.3.0