eclass/java-utils-2.eclass: Implemented default EAPI=6 src_prepare
Duplicated code from EAPI=6 default src_prepare, rather call existing

This should allow java ebuilds to use EAPI=6, but does not mean the
eclass fully supports all EAPI=6 changes. The lack of such caused every
ebuild using EAPI=6 to fail without calling eapply_user in src_prepare.
Ebuilds calling epatch or using PATCHES need to be modified when
updating to EAPI=6 or this will cause issues as patches will be applied
twice and fail the 2nd time. If not the first with eapply, as it
requires -p1 and some patchs might not be.