x11-libs/xcb-util-cursor: use gentoo-specific cursor paths from libXcursor
Gentoo has a long strange tradition of using /usr/share/cursors/xorg-x11
for cursors. Love it or hate it, that's how it is. x11-libs/libXcursor has
respected this. However, until recently -- commit here [1] -- xcb-util-cursor
did not support this. Now, finally, it does. So, we copy over the same paths
as with libXcursor. This fixes numerous bugs.

Package-Manager: portage-2.3.0