dev-lang/perl: Update upgrade prose to encourage depclean
Depcleans really do help a lot and you should do them lots.

Note: the 5.24.0-r1 version ( which is patched here ) is not displayed
to the user due to some mistakes made when flattening it, this will be
rectified in a future commit.

Note: Patching apparently "Old" eblits because due to a lack of clarity
on how eblits actually work, it turns out that all eblit-using
dev-lang/perl ebuilds are sourcing *BOTH* their own eblit entries *AND*
the eblit for v5016001!!.

The use of eblits here has made it impossible to clearly discern which
code is actually running, and that perl 5.22* and 5.20* actually work
at all is more a case of amazing amounts of good luck than actual
planning.

Package-Manager: portage-2.3.0