dev-lang/perl: Restore missing call to check_rebuild
In the process of flattening perl-5.24.0-r1.ebuild, the call
to check_rebuild was omitted unintentionally, removing the ability
for this message to be displayed.

This commit reinserts the displaying of said message at the pkg_preinst
phase the last possible moment before the system perl gets updated for real.

Package-Manager: portage-2.3.0