profiles: update package.use.mask for net-dns/dnsmasq[conntrack]
net-libs/libnetfilter_conntrack is stable on alpha, arm, mips and
ppc; so no masking is needed on these archs. It is also unecessary
for nios since neither dnsmasq nor libnetfilter_conntrack have
keywords there. On the other hand, a mask is required for the
FreeBSD and mips profiles since there dnsmasq is keyworded while
libnetfilter_conntrack is not.

Gentoo-Bug: https://bugs.gentoo.org/585542