media-plugins/gst-plugins-opus: Properly split again
... instead of building as part of gst-plugins-base[opus].
This makes gst-plugins-opus a split of -base instead of -bad as it
was moved from -base to -bad with some leftovers in -bad.
The ext/opus/ leftover part in gst-plugins-bad is not necessary
for anything and we can safely not package that at all.
This should give a nicer upgrade path for stable opus plugin users.

Gentoo-bug: 589440