licenses: add new Zend-2.0 license as part of the @MISC-FREE group.
The "Zend Engine License, version 2.00" covers parts of the PHP
runtime. The latest version of the Zend license is vanilla BSD:

  https://framework.zend.com/license

However, all of the sources shipped with dev-lang/php still refer to
version 2.00 of the license, so we're including it. Version 2.00 is a
modified BSD license with a few more attribution clauses thrown in. It
is still a free software license (although not approved, to my
knowledge), so it has been added to the @MISC-FREE license group.

Gentoo-Bug: 595664