dev-lang/php: new 5.6 and 7.0 revisions with additional LICENSE entries.
The main license under which PHP is distributed is the "PHP license."
But, a lot of files, functions, and libraries have been inlined and
bring with them the associated license from the upstream project. This
commit adds the following unconditionally:

  * BSD (thread-safe resource manager, probably other places)
  * Zend-2.0 (used by the Zend engine, everything under ${S}/Zend/)

and the following USE-conditionally:

  * LGPL-2.1+, used by the bcmath extension.
  * BSD-2, used by the fpm SAPI and mbstring extension.
  * gd, used by gd.
  * LGPL-2.1, used by the mbstring extension.

The main license was also updated from "PHP-3" to "PHP-3.01", because
that's what it is.

Package-Manager: portage-2.3.0