x11-libs/libxcb: Revert USE=doc controlling man pages
FEATURES=noman is for skipping man pages, not doc USE flag.
This change seems to not have been acknowledge by any libxcb maintainers.
It may be discussed on bugzilla on the premise of these being development
docs, some of which are already behind USE=doc due to extra deps.

Package-Manager: portage-2.3.1