x11-libs/libxcb: Lock subslot to 1.12 for now
As it was before, a simple bugfix release of 1.12.1 would force
a huge set of rebuilds unnecessarily. We'll need to decide if we
need subslot 1.13 upon such a release appearing or not.
This is a revbump due to a -r1 having just been reverted and so
avoids it looking like a downgrade.

Gentoo-bug: 576890
Package-Manager: portage-2.3.1