dev-lang/php: preserve timestamps when creating SAPI build directories.
We create several copies of ${S} -- one for each SAPI -- during the
build. Until now, those copies have been made with "cp -r". I believe
this caused unnecessary regeneration of the Zend parsers, opening the
door for version mismatches of sys-devel/bison.

For example, the file

  ${S}/Zend/zend_language_parser.c

is generated from

  ${S}/Zend/zend_language_parser.y,

but the ".c" file should already exist in the source tree. In the
upstream tarball, the former always exists (pre-generated) and has a
timestamp later than the latter. This prevents the build system from
attempting to call "yacc" on zend_language_parser.y. When we use
"cp -r" to copy the source tree, I find timestamps like the following:

  2016-08-16 19:08 zend_language_parser.c
  2016-08-16 19:08 zend_language_parser.y

They look identical, but if the copy was performed in alphabetical
order, I suspect that the ".y" file looks newer than the ".c" file
that it generates. It is therefore my educated guess that the build
system is using bison to regenerate those ".c" files when in fact
there is no need to do so. By using "cp -a" instead of "cp -r", I
hope to preserve the ages of those files, and eliminate bison from
normal end-user builds of dev-lang/php.

The fix has been applied to the latest 5.6 and 7.0 revisions in-place.

Gentoo-Bug: 595792

Package-Manager: portage-2.3.0