dev-python/mimeparse: use the actual ABI version
In src_install there was a bit of scripted logic to determine
the ABI used for link creation. This wasn't working with a
newer version of pypy3. Instead of the logic a small Python
script is implemented that should always get the correct
ABI version.

Gentoo-Bug: https://bugs.gentoo.org/show_bug.cgi?id=586008

Package-Manager: portage-2.3.0