net-irc/hexchat: fix libgdiplus.so & Werror errors
Fixes the System.DllNotFounException: libgdiplus.so error when the
theme-manager USE flag is set.  The problem is that only mono[minimal]
is buildable right now, so mono[-minimal] does not work.  However,
mono[minimal] does not pull in libgdiplus.so.  Therefore, the "fix" is
to pull in libgdiplus as a dependency if the user is using
mono[minimal].  Unfortunately, this doesn't fix the mono[-minimal] use
case.

Additionally, removes all Werror* flags from configure.ac.

Gentoo-bug: 578290
Gentoo-bug: 597126

Package-Manager: portage-2.3.2
Closes: https://github.com/gentoo/gentoo/pull/2563

Signed-off-by: David Seifert <soap@gentoo.org>