net-misc/netifrc: bump, MANY fixes.
This release of netifrc should hopefully represent the next stable
candidate. The 0.3.x and 0.4.x series had some bugs that should not be
in stable.

One specific change to note, is that if an address or route is already
configured on an interface, prior to starting the interface service, the
interface will now start successfully, whereas it would previously fail.
If the parameters on the existing route and new route are slightly
different, this may lead to non-deterministic behavior.

Bugs fixed:
420941, 588872: l2tp/dummy/hsr support.
487208: support busybox better, no scope=host on loopback.
488250, 488260, 488262: better IPv6 address handling.
489398: more reliable interface startup
507346: resolvconf-gentoo renamed to openresolv
524156: DHCP classless static routes for udhcpc
525958: Better POSIX sh support
540728, 547906: macchanger 1.7 support
545364: Switch to /run instead of /var/run
575128: newer udev support
596234: New iproute keywords: noprefixroute mngtmpaddr
596844: metric_SSID support

Package-Manager: portage-2.3.2
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>