sys-fs/e2fsprogs: patch mke2fs.conf.in #598100
A commit prior to e2fsprogs 1.43 release accidentally removed a
mke2fs.conf flag and resulted in mke2fs becoming very slow (takes 5
minutes to format a 1.5TB disk as opposed to <1 minute). This just pulls
in the upstream patch.

More details and the original commit message can be found in the patch
file itself.

Gentoo-Bug: 598100