dev-db/postgresql: Fix broken/improve no-server patches
9.4 and 9.5 had a small change to a section that is modified by the
patch, thus causing failure.

Further, pgbench was added back in as it is a client program that does
not need the server to be on the same machine.

Lastly, the 9.6 no-server patch was incomplete and allowed the entire
backend to be built when only the headers are needed…theoretically.

Package-Manager: portage-2.2.28