sci-misc/boinc: Add suspend/resume to boinc init script, fix null byte input.
Gentoo-Bug: 493476

Enable users to suspend/resume all projects without having to start
and use the manager GUI.

Gentoo-Bug: 584386 (partly)

After upgrading to the current app-shells/bash-4.4 the stop command
provokes the following warning:

"command substitution: ignored null byte in input"

This happens due to the usage of "cut" with --output-delimiter=''.
Obviously "cut" puts a null byte out if the delimiter was set to
nothing. (Checked with hexdump, it does.)

The fix is to use "tr -d ." to concatenate the version numbers, and
then to compare against an "expr substr".

Package-Manager: portage-2.3.2