net-libs/nodejs: add optional support for dtrace-based tracing
NodeJS supports dtrace-based tracing, that can be used by
dev-util/systemtap and, more recently, dev-util/bcc. This support is
forcibly disabled in nodejs ebuilds.

This commit adds a systemtap USE flag to be able to turn on this
support conditionally. The name of this flag matches the choice done
for other ebuilds, like dev-db/mysql or app-emulation/qemu. Despite
what has been done in those ebuilds, however, a dependency on
dev-util/systemtap is not added: this is because, as of late,
systemtap is not the only consumer of those traces anymore.