dev-php/pecl-http: remove old masked versions.
This removal is in response to CVE-2016-5873. That vulnerability may
not in fact affect our pecl-http-1.x, according to Comment 2 on the
associated bug. However, php-5.5 is end-of-life and masked in the
tree, and the removed versions of pecl-http only support php-5.5. As
a result, their days were numbered anyway.

Gentoo-Bug: 587466

Package-Manager: portage-2.3.0