cuda.eclass: fixes gcc-bindir -f
Currently the check against *g++* doesn't work, thus the NVCCFLAGS
are never set accordingly.
The check for gnu is now made through tc-is-gcc.

Also the '=' with the "" around the bindir are causing a "not a valid
directory" error while compiling with cuda support.

It now works as expected:
	with gcc-5.4.0 and cuda 7.5 NVCCFLAGS are set properly to
gcc-4.9.x