sci-misc/boinc: Rename desktop file to boincmgr.desktop
Gentoo-Bug: 599910

This is a workaround to the following issue:
The desktop entry filename and the executable filename are different.
Binary:       boincmgr
Desktop file: boincmgr-boinc.desktop
This causes GNOME to use an ugly icon in the launcher. It is also not
possible to add the BOINC icon to the favorites.
Renaming the desktop file to boincmgr.desktop fixes both issues,
although this is only a hotfix. The GNOME shell should be perfectly
capable of using any well formed desktop file regardless of its name.

Package-Manager: portage-2.3.3