sys-devel/gettext: force a recent version of libxml2 #596918
Gettext, via gnulib, bundles libxml2.  If it detects the system version
of libxml2 is too old, it will automatically fall back to that.  Pass
configure flags to disable that logic, but also add a min version to
the libxml2 dep to match the versions that are included.

Both of these include libxml2-2.9.3, so that's what we force.  We'll
have to remember to check the version of the bundled copy from time
to time in case we need to update the dependency.