dev-lang/php: three new revisions adding ACL support for php-fpm.
The php-fpm backend can be used to host multiple applications with
different settings and permissions. Traditionally, permissions on its
UNIX socket are granted to the "listen.owner" and "listen.group"
defined in php-fpm.conf. However, php-fpm can be compiled with POSIX
ACL support, after which the more-powerful "listen.acl_users" and
"listen.acl_groups" can be used.

This commit adds three new revisions -- one for each supported branch
of PHP. To each new revision, the "acl" USE flag has been added. When
USE=acl is set, sys-apps/acl gets pulled in as a dependency, and
"--with-fpm-acl" is passed to the configure script.

Gentoo-Bug: 601634

Package-Manager: portage-2.3.0