dev-lang/php: allow app-text/htmltidy in php:7.1.
To help phase out the abandoned and buggy app-text/htmltidy, I tried
to make the new php:7.1 depend on app-text/tidy-html5
instead. However, since older versions of PHP do not support
tidy-html5, that causes a problem: php:7.1 cannot be installed
alongside the older versions when USE=tidy is set for both of them.

This solution is somewhat of a cop out, allowing htmltidy to be used
in php:7.1 as a last resort. The alternative would have been to
backport the tidy-html5 support to (both) older versions of PHP. That
may eventually become necessary if it looks like htmltidy will be
masked for removal; however, for now, this is the path of least
resistance.

Gentoo-Bug: 601614

Package-Manager: portage-2.3.0