net-analyzer/nagios-core: remove proxy maintainers from the metadata.
We haven't heard from this package's proxy maintainer in a while, but
new bugs are still being assigned to him. As long as I'm the de-facto
maintainer, I would prefer that they be assigned to me. This commit
removes both the proxied and proxy maintainers, as well as the
proxy-maint project (who gave the OK for this removal).

Package-Manager: portage-2.3.0