dev-php/php-redmine-api: stabilize php-redmine-api-1.5.9 on x86.
Maintainer stabilization: I have the hardware and am able to test
this. It does what it's supposed to, and since the code is pure PHP, I
probably should have filed the initial stabilization request with the
ALLARCHES keyword. Nevertheless... the stabilization request was opened
over three months ago, so I think it's past the "timeout" period.

Gentoo-Bug: 592262

Package-Manager: portage-2.3.0