sys-apps/apply-default-acl: stabilize v0.0.6 on x86 and amd64.
Maintainer stabilization: the stabilization request for this was only
filed about a month ago, but the new version contains an important fix
to symlink handling that should go stable as soon as possible. (Contrary
to the documentation, symlinks *were* followed in v0.0.4.)

I regularly use this package on both x86 and amd64, and it passes the
test suite on both architectures.

Gentoo-Bug: 599172

Package-Manager: portage-2.3.0