net-firewall/nftables: add xml USE flag
libnftnl-1.0.7 no longer supports outputing the firewall state as XML.
If a user wants that output, the user must use libnftnl-1.0.6.
Therefore, I added the xml USE flag so I could restrict libnftnl to
1.0.6 when xml is enabled and allow newer versions to be used when it is
not.

Unforunately, nftables unconditionally compiles support for xml output,
so when xml use is attempted with a version of libnftnl that does not
support it, the result is just an empty file.

Package-Manager: Portage-2.3.3, Repoman-2.3.1