dev-php/twig: add twig-1.29.0 with support for building the C extension.
A few versions of Twig have been released since the version we have in
the tree, so the first goal of the new ebuild is simply to bring our
tree to parity with upstream.

But the new ebuild also adds a feature, the ability to build the C
extension for Twig. In older versions of PHP, the C implementation
provides improved performance over the pure-PHP implementation. The
only "older version" we support is php:5.6, so the C extension is only
built when the user's PHP_TARGETS contains "php5-6". The extension
will only be built with USE=extension; otherwise users who did not
want the extension would be required to add "php5-6" to PHP_TARGETS in
order to appease our eclass.

Upstream does not currently support building the extension against
php:7.0 or php:7.1, but the performance improvement is not so great
with those versions of PHP anyway.

Gentoo-Bug: 566468
Gentoo-Bug: 566470
Suggested-by: Dennis Schridde

Package-Manager: portage-2.3.0