dev-php/PHPMailer: new version to fix CVE-2016-10033.
The new version fixes a critical security vulnerability, and so it
replaces the existing (unstable) version. The release tarball for the
latest version is missing some documentation. I reported that
upstream, and commented those docs out of the ebuild for now so that
we can get the CVE fix as soon as possible.

Gentoo-Bug: 603750
PHPMailer-Bug: https://github.com/PHPMailer/PHPMailer/issues/919

Package-Manager: portage-2.3.0