apache-2.eclass: drop path to --with-ssl #516240
The apache configure script assumes that, if you passed in a path, that
it's to a standalone install of openssl.  It'll then go poking in that
dir directly and possible add flags for building against it.  This also
breaks building/linking of other libraries, in both multilib and cross-
compile setups.

By dropping the path, we force apache to do the lookup via pkg-config.